Move rate limit circumvention to generic
